requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sentence-transformers>=2.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.9
pydantic>=2.0.0
torch>=2.0.0
